+++ app/src/main/kotlin/me/xizzhu/android/joshua/search/SearchModule.kt
@@ -24,9 +24,8 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityComponent
 import dagger.hilt.android.scopes.ActivityScoped
-import me.xizzhu.android.joshua.core.BibleReadingManager
 import me.xizzhu.android.joshua.Navigator
-import me.xizzhu.android.joshua.core.SettingsManager
+import me.xizzhu.android.joshua.core.*
 import me.xizzhu.android.joshua.search.result.SearchResultListPresenter
 import me.xizzhu.android.joshua.search.toolbar.SearchToolbarPresenter
 
@@ -52,11 +51,14 @@ object SearchModule {
     @Provides
     fun provideSearchViewModel(searchActivity: SearchActivity,
                                bibleReadingManager: BibleReadingManager,
+                               bookmarkManager: VerseAnnotationManager<Bookmark>,
+                               highlightManager: VerseAnnotationManager<Highlight>,
+                               noteManager: VerseAnnotationManager<Note>,
                                settingsManager: SettingsManager): SearchViewModel {
         val factory = object : ViewModelProvider.Factory {
             override fun <T : ViewModel> create(modelClass: Class<T>): T {
                 if (modelClass.isAssignableFrom(SearchViewModel::class.java)) {
-                    return SearchViewModel(bibleReadingManager, settingsManager) as T
+                    return SearchViewModel(bibleReadingManager, bookmarkManager, highlightManager, noteManager, settingsManager) as T
                 }
 
                 throw IllegalArgumentException("Unsupported model class - $modelClass")