+++ app/src/test/kotlin/me/xizzhu/android/joshua/search/result/SearchResultListPresenterTest.kt
@@ -24,8 +24,8 @@ import kotlinx.coroutines.flow.flow
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.runBlocking
 import me.xizzhu.android.joshua.Navigator
-import me.xizzhu.android.joshua.core.Settings
-import me.xizzhu.android.joshua.core.VerseIndex
+import me.xizzhu.android.joshua.R
+import me.xizzhu.android.joshua.core.*
 import me.xizzhu.android.joshua.search.SearchActivity
 import me.xizzhu.android.joshua.search.SearchRequest
 import me.xizzhu.android.joshua.search.SearchResult
@@ -44,14 +44,19 @@ import kotlin.test.assertEquals
 class SearchResultListPresenterTest : BaseUnitTest() {
     @Mock
     private lateinit var lifecycle: Lifecycle
+
     @Mock
     private lateinit var navigator: Navigator
+
     @Mock
     private lateinit var searchViewModel: SearchViewModel
+
     @Mock
     private lateinit var searchActivity: SearchActivity
+
     @Mock
     private lateinit var loadingSpinner: ProgressBar
+
     @Mock
     private lateinit var searchResultListView: CommonRecyclerView
 
@@ -63,6 +68,9 @@ class SearchResultListPresenterTest : BaseUnitTest() {
         super.setup()
 
         `when`(searchActivity.lifecycle).thenReturn(lifecycle)
+        `when`(searchActivity.getString(R.string.title_bookmarks)).thenReturn("Bookmarks")
+        `when`(searchActivity.getString(R.string.title_highlights)).thenReturn("Highlights")
+        `when`(searchActivity.getString(R.string.title_notes)).thenReturn("Notes")
         `when`(searchViewModel.settings()).thenReturn(emptyFlow())
         `when`(searchViewModel.searchRequest).thenReturn(emptyFlow())
 
@@ -83,9 +91,9 @@ class SearchResultListPresenterTest : BaseUnitTest() {
     @Test
     fun testObserveSearchRequest() = runBlocking {
         val query = "query"
-        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, false)))
-        `when`(searchViewModel.search(query))
-                .thenReturn(flowOf(SearchResult(query, emptyList(), emptyList(), emptyList())))
+        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, false, true, false, true)))
+        `when`(searchViewModel.search(query, true, false, true))
+                .thenReturn(flowOf(SearchResult(query, emptyList(), emptyList(), emptyList(), emptyList(), emptyList(), emptyList())))
 
         searchResultListPresenter.onCreate()
 
@@ -105,9 +113,9 @@ class SearchResultListPresenterTest : BaseUnitTest() {
     @Test
     fun testObserveSearchRequestWithInstantSearch() = runBlocking {
         val query = "query"
-        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, true)))
-        `when`(searchViewModel.search(query))
-                .thenReturn(flowOf(SearchResult(query, emptyList(), emptyList(), emptyList())))
+        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, true, false, false, true)))
+        `when`(searchViewModel.search(query, false, false, true))
+                .thenReturn(flowOf(SearchResult(query, emptyList(), emptyList(), emptyList(), emptyList(), emptyList(), emptyList())))
 
         searchResultListPresenter.onCreate()
 
@@ -123,8 +131,8 @@ class SearchResultListPresenterTest : BaseUnitTest() {
     @Test
     fun testObserveSearchRequestWithException() = runBlocking {
         val query = "query"
-        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, false)))
-        `when`(searchViewModel.search(query)).thenReturn(flow { throw RuntimeException() })
+        `when`(searchViewModel.searchRequest).thenReturn(flowOf(SearchRequest(query, false, true, true, true)))
+        `when`(searchViewModel.search(query, true, true, true)).thenReturn(flow { throw RuntimeException() })
 
         searchResultListPresenter.onCreate()
 
@@ -143,15 +151,30 @@ class SearchResultListPresenterTest : BaseUnitTest() {
     fun testToItems() {
         val query = "query"
         val expected = listOf(
+                TitleItem("Notes", false),
+                SearchNoteItem(VerseIndex(0, 0, 3), MockContents.kjvBookShortNames[0], MockContents.kjvVerses[3].text.text, "note", query, searchResultListPresenter::selectVerse),
+                TitleItem("Bookmarks", false),
+                SearchVerseItem(
+                        VerseIndex(0, 0, 1), MockContents.kjvBookShortNames[0],
+                        MockContents.kjvVerses[1].text.text, query, Highlight.COLOR_NONE, searchResultListPresenter::selectVerse
+                ),
+                TitleItem("Highlights", false),
+                SearchVerseItem(
+                        VerseIndex(0, 0, 2), MockContents.kjvBookShortNames[0],
+                        MockContents.kjvVerses[2].text.text, query, Highlight.COLOR_BLUE, searchResultListPresenter::selectVerse
+                ),
                 TitleItem(MockContents.kjvBookNames[0], false),
-                SearchItem(
+                SearchVerseItem(
                         VerseIndex(0, 0, 0), MockContents.kjvBookShortNames[0],
-                        MockContents.kjvVerses[0].text.text, query, searchResultListPresenter::selectVerse
+                        MockContents.kjvVerses[0].text.text, query, Highlight.COLOR_NONE, searchResultListPresenter::selectVerse
                 )
         )
 
         val searchResult = SearchResult(
                 query, listOf(MockContents.kjvVerses[0]),
+                listOf(Pair(Bookmark(VerseIndex(0, 0, 1), 0L), MockContents.kjvVerses[1])),
+                listOf(Pair(Highlight(VerseIndex(0, 0, 2), Highlight.COLOR_BLUE, 0L), MockContents.kjvVerses[2])),
+                listOf(Pair(Note(VerseIndex(0, 0, 3), "note", 0L), MockContents.kjvVerses[3])),
                 MockContents.kjvBookNames, MockContents.kjvBookShortNames
         )
         val actual = with(searchResultListPresenter) { searchResult.toItems() }